Split base_program.py into: - main.py - program entry point - team.py - team data + operations - utils.py - helper functions
