+++ tests/test_main.py
@@ -2,7 +2,7 @@ import sys
 import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src')))
-from base_program import format_greeting, count_team_members
+from main import format_greeting, count_team_members
 
 
 def run_tests():
